\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,pssm-method}
\title{Methods for Function \code{summary} in Package \pkg{pssm}}
\description{
Methods for function \code{summary} in package \pkg{base}
}
\section{Methods}{
\describe{
\item{\code{signature(object = "pssm")}}{
Produces a \code{pssm.summary} object with the coefficients of the covariates, their standard erros and confidence bounds in a data frame.
}
}}
\keyword{methods}

